.claude/skills/job-application-tailor/requirements.txt
python-docx>=1.1.2
pyyaml>=6.0.2
jsonschema>=4.21.0
docx2pdf>=0.1.8
